requirements.txt
aiohttp==3.6.2
alembic==1.4.2
apply-defaults==0.1.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
json-rpc==1.13.0
jsonrpcclient==3.3.6
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
msal==1.4.3
multidict==4.7.6
pycparser==2.20
PyJWT==1.4.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
regex==2020.10.28
requests==2.24.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.18
typing-extensions==3.7.4.2
urllib3==1.25.10
# uWSGI==2.0.19.1
Werkzeug==1.0.1
yarl==1.5.1
Flask-JWT==0.3.2
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pyModbusTCP==0.1.8
toml==0.10.1
wrapt==1.12.1
astroid==2.4.2
psycopg2-binary==2.8.5